testing a endgame for the incoming runes. I does not look half bad...
I suspect (after testing it a bit) I will need to optimise the rendering
of the scars/scratches in the shard. Instead of keeping a loty of
objects I should simplify the paths and add them to an array of
"scratches" I can go over in each iteration.
And delete or reuse the actual rune objects.